requirements.txt
fastapi
uvicorn
phidata
selenium
playwright
beautifulsoup4
requests
lxml
python-dotenv
